backend/requirements.txt
fastapi
uvicorn
pymongo
motor
pydantic
passlib
bcrypt
python-jose
python-dotenv
jq

